Replace occurrences of long with int
This ensures that integer widths are as expected for 64-bit builds that use the LP64 data model (i.e. on *nix platforms).

+++ doomclassic/doom/i_net.cpp
@@ -83,7 +83,7 @@ namespace {
 int	DOOMPORT = 1002;	// DHM - Nerve :: On original XBox, ports 1000 - 1255 saved you a byte on every packet.  360 too?
 
 
-unsigned long GetServerIP() {
+unsigned int GetServerIP() {
 	return ::g->sendaddress[::g->doomcom.consoleplayer].sin_addr.s_addr;
 }
