+++ neo/idlib/Parser.cpp
@@ -1318,7 +1318,7 @@ typedef struct operator_s
 
 typedef struct value_s
 {
-	signed long int intvalue;
+	signed int intvalue;
 	double floatvalue;
 	int parentheses;
 	struct value_s *prev, *next;
@@ -1389,7 +1389,7 @@ int PC_OperatorPriority(int op) {
 
 #define FreeOperator(op)
 
-int idParser::EvaluateTokens( idToken *tokens, signed long int *intvalue, double *floatvalue, int integer ) {
+int idParser::EvaluateTokens( idToken *tokens, signed int *intvalue, double *floatvalue, int integer ) {
 	operator_t *o, *firstoperator, *lastoperator;
 	value_t *v, *firstvalue, *lastvalue, *v1, *v2;
 	idToken *t;
@@ -1814,7 +1814,7 @@ int idParser::EvaluateTokens( idToken *tokens, signed long int *intvalue, double
 idParser::Evaluate
 ================
 */
-int idParser::Evaluate( signed long int *intvalue, double *floatvalue, int integer ) {
+int idParser::Evaluate( signed int *intvalue, double *floatvalue, int integer ) {
 	idToken token, *firsttoken, *lasttoken;
 	idToken *t, *nexttoken;
 	define_t *define;
@@ -1905,7 +1905,7 @@ int idParser::Evaluate( signed long int *intvalue, double *floatvalue, int integ
 idParser::DollarEvaluate
 ================
 */
-int idParser::DollarEvaluate( signed long int *intvalue, double *floatvalue, int integer) {
+int idParser::DollarEvaluate( signed int *intvalue, double *floatvalue, int integer) {
 	int indent, defined = false;
 	idToken token, *firsttoken, *lasttoken;
 	idToken *t, *nexttoken;
@@ -2007,7 +2007,7 @@ idParser::Directive_elif
 ================
 */
 int idParser::Directive_elif() {
-	signed long int value;
+	signed int value;
 	int type, skip;
 
 	idParser::PopIndent( &type, &skip );
@@ -2029,7 +2029,7 @@ idParser::Directive_if
 ================
 */
 int idParser::Directive_if() {
-	signed long int value;
+	signed int value;
 	int skip;
 
 	if ( !idParser::Evaluate( &value, NULL, true ) ) {
@@ -2125,7 +2125,7 @@ idParser::Directive_eval
 ================
 */
 int idParser::Directive_eval() {
-	signed long int value;
+	signed int value;
 	idToken token;
 	char buf[128];
 
@@ -2264,7 +2264,7 @@ idParser::DollarDirective_evalint
 ================
 */
 int idParser::DollarDirective_evalint() {
-	signed long int value;
+	signed int value;
 	idToken token;
 	char buf[128];
 
@@ -2313,7 +2313,7 @@ int idParser::DollarDirective_evalfloat() {
 	token = buf;
 	token.type = TT_NUMBER;
 	token.subtype = TT_FLOAT | TT_LONG | TT_DECIMAL | TT_VALUESVALID;
-	token.intvalue = (unsigned long) fabs( value );
+	token.intvalue = (unsigned int) fabs( value );
 	token.floatvalue = fabs( value );
 	idParser::UnreadSourceToken( &token );
 	if ( value < 0 ) {
